Переделываю ObjectSelector под double

+++ route_editor/include/ObjectSelector.h
@@ -51,8 +51,8 @@ private:
 
     vsg::dvec3 prev_intersect_pos;
     vsg::dvec3 total_translation;
-    float total_rotation_rad;
-    vsg::vec3 total_scale;
+    double total_rotation_rad;
+    vsg::dvec3 total_scale;
     vsg::ref_ptr<SingleSwitch> front_plane_switch;
     vsg::vec3 front_plane_up;
 };
